refactor(folio): refactor card folio wokrflow into folio controller
